authz: parse and validate the JWT, make available through Env
the idea is to parse and validate the token only once, and pass it
on to lower layers for whatever additional inspection.

this came up when the verification logic in useradm was considered.
it doesn't make sense to decode the token at useradm.Verify(), because
we just did it an upper layer - the only problem is that the result was not
exposed.

Issues: MEN-798
Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>