authz_simple: skip auth rules checking on POST /verify
this causes 403 errors on pretty much any management API call in the system.
note that all these requests are routed to /verify first - which subsequently
denies access since the authorizer tries to check the request against useradm's
own authz rules.

the quick solution is to make the authorizer aware of this, and return asap on POST /verify.
effectively, the actual authz will be delegated to the target downstream service;
useradm will only check the token validity, as it should.

Issues: MEN-907

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>