middleware: fire authz middleware only on /verify
since all requests go through /verify, it effectively becomes a
centralized authz endpoint, having all the data on allowed scopes
vs actions vs services in the system. thus, it is the only endpoint
actually interested in having the middleware enabled.

at this point, the middleware is kind of spurious - the relevant checks
could simply be moved to the /verify handler. as it stands, the mw can not
be reused by other services and could be removed, but I'm leaving it in just to
avoid a larger rework. further refactors and redesign can be pushed post 1.0, when
we actually include 'real' authorization.

Issues: MEN-918

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>