user/useradm: remove 'initial user' login logic
from now on, the user must be created in the database
e.g. via the create-user cli. we won't support the initial
login flow, which allowed obtaining temp credentials for
creating the initial user (if the db was empty).

doInitialLogin is out, doRegularLogin becomes simply Login.

Issues: MEN-1034

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>