jwt/token: simplify token handling
we have disparate ways of signing/parsing the tokens, and it's
slightly confusing.

one impl. is the Marshal/Unmarshal pair in jwt/token.go:Token, the
other is ToJWT/FromJWT in jwt/jwt.go:JWTHandler. the former in fact delegates
to the latter. what's more, Unmarshal is actually undefined and never used.

it makes sense to simplify this, and the reasonable way is to get rid of
the Marshal/Unmarshal pair - the JWTHandler wins since:
- it's already used in the authz middleware (would be hacky to pass a
full useradm instance there with it's funcs just to handle the token)
- it already has everything we need

Issues: MEN-1063

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>