main, main_test: add testing function for running acceptance tests
this is the 'main' that will be called during acceptance tests.

note a workaround for flags vs urfave/cli conflicts:
during tests, we have to pass several golang flags, incompatible with cli;
all cli-specific arguments will then be passed as the -cli-args flag.
to enable this behavior, the main() logic is factored out to a parametrized
doMain(args) func.

this is not strictly necessary for now, but might become handy in case we want to
pass custom configs/params. the pattern is used in all new services/tests.

Issues: MEN-1223

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>