user: most of CreateUser logic moved to a private func
CreateUser is meant strictly for the mgmt api, which always
hashes and always propagates the incoming data.

a new function will be introduced for the internal api, operating on
its model. most of the creation logic is therefore moved to a common
func to be reused by the two methods.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>